backend/requirements.txt

Flask==2.3.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
Flask-Migrate==4.0.4
marshmallow==3.19.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
pytest==7.4.0
pytest-flask==1.2.0
Werkzeug==2.3.7
redis==4.5.4
celery==5.3.1
gunicorn==21.2.0
qrcode==7.4.2
Pillow==10.0.0